2. 反思设计模式[Reflection Design Pattern]/2.7 无评分实验-用反思改进SQL生成[Ungraded Lab- Improving SQL Generation with Reflection]/requirements.txt
aisuite
python-dotenv
pandas
